+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - The `phenocycler` reader can now also read `.tif` files (not just `.qptiff`)
 - Added missing legend in the HTML report under the "Channels" section (#15)
+- The cell area is also stored in the table (in `.obs["area"]`)
 
 ## [1.0.2] - 2024-01-15